services/research/experimenting/requirements.txt
opencv-python
pillow
torch
torchvision
torchaudio
pandas
psutil
PyYAML
tqdm
ultralytics
gitpython